予約変更機能を追加

+++ rese-nuxtpj/components/BookingInfo.vue
@@ -2,7 +2,7 @@
   <div>
     <section class="booking_card" :booking="booking" :index="index">
       <span>予約{{index+1}}</span>
-      <img @click="cancelBooking(booking.id)" class="close-btn" src="~assets/images/close.png">
+      <img @click="cancelBooking(booking.id, index+1)" class="close-btn" src="~assets/images/close.png">
         <table>
           <tr>
             <th>Shop</th>
@@ -21,7 +21,8 @@
             <td>{{booking.number_of_people}}人</td>
           </tr>
         </table>
-      </section>
+        <NuxtLink class="change-booking" :to="{path:'bookings/' + booking.id}">日時、人数を変更する</NuxtLink>
+    </section>
   </div>
 </template>
 
@@ -41,9 +42,11 @@ export default {
     }
   },
   methods:{
-    async cancelBooking(id){
-      await this.$axios.delete('api/booking/' + id);
-      await this.$store.dispatch('getMyBookings');
+    async cancelBooking(id, index){
+      if(confirm('予約'+index +'を削除してよろしいですか？')){
+        await this.$axios.delete('api/booking/' + id);
+        await this.$store.dispatch('getMyBookings');
+      }
     }
   }
 }
@@ -70,64 +73,18 @@ export default {
   }
   .booking_card th{
     text-align:left;
-    padding-right:20px;
+    padding:5px 20px 5px 0;
   }
 
-  .favorite-shop_cards{
-    display:flex;
-    flex-wrap:wrap;
-    position:relative;
-  }
-  .card{
-    border-radius:5px;
-    overflow:hidden;
-    background-color:white;
-    box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.3);
-    margin:10px 5px;
-    width:45%;
-    position:relative;
-  }
-  .card-content{
-    padding:10px 20px;
-  }
-  .favorite-shops{
-    width:50%;
-  }
-  .card-img{
-    max-width:100%;
-    height:auto;
-  }
-  .card-name{
-    font-weight:bold;
-    font-size:18px;
-    margin: 10px 0;
-  }
-  .card-content span{
-    font-size:15px;
-  }
-  .card-detail{
-    display:inline-block;
-    width:100px;
-    font-size:15px;
-    padding:7px 5px;
-    margin-top:20px;
-    text-align:center;
-    text-decoration:none;
-    background:royalblue;
-    color:white;
-    border-radius:5px;
-  }
-  .fav-btn{
-    width:35px;
-    display:inline-block;
-    position:absolute;
-    bottom:5;
-    right:5;
-  }
   .close-btn{
     height:23px;
     position:absolute;
     top:10%;
     right:5%;
+  }
+  .change-booking{
+    display: block;
+    color: white;
+    padding:10px;
   }
 </style>
